Changed Navigation and Landing Component
Styled the Navigation component to be smaller with smaller text. The menu button is now a hamburger menu. The position is fixed. The Landing component has a single image with text cover it. Added media queries to make the image responsive. Added functionality to change the background image based on the string inside backgroundImages array