Made major changes to layout
Made changes to layout and styling of several components: Navigation, Landing, About. Navigation is fixed to top. Landing page is separated into two sections. Left side has an inner image. In About Component there is a text describing what the purpose of the app is, and below that is an image that also has a inner image.