Feat: Add animation to hamburger icon that will call Menu component
Refactors the Navigation component. The Hamburger menu is made up of two divs that are animated. Passing the handleShowMenu method to Menu component and Navigation to toggle the Menu. Using gsap to animate the hamburger icon, which then toggles Menu into view. Adds a timeout function which return a promisified setTimeout function.