Feat: Add global user object that is set from login
Add getUser method that will retreive a user from localStorage. The method acts as the default for user, a new state variable. Add a loginUser method which will set a user to local storage if there isn't a user already. The loginUser method is the value for the AuthContext.provider's login property, and the user state variable is the value for AuthContext's user. The Menu dynamically will render a link based on the logged in user. The Login component calls AuthContext login method to set the global user variable to the logged in user.