Refactor: Replace useToken hook with useAuth custom hook
The useToken has been renamed to useAuth because it handles not only token state, but user state as well. The useAuth hook returns login and logout methods which handles storing and removing user and token from local storage. useAuth also retuns the token and user. The authContext.Provider uses these as values for its properties. Replaces logic from components that previously used useToken and updates it to use useAuth. The components affected by this change are: App, Login, and Signup.