Feat: Add SearchModal and Articles component to AllArticlesPage.js
AllArticlesPage component renders a short summary of the purpose of the page and a dynamic grid of articles. SearchModal is an input that changes the state of AllArticlesPage component. The user types in some text, and relevant articles will be shown. If no text is inputted, all articles are shown.