Feat: Add /me route, that renders page according to user role
Adds new Route in App component, which renders component based on the role of the logged in user. Add the correct path for the link in the Menu component. Creates an Author Component that renders many children components. Author contains a sidebar, main, and profile section. Routes are dynamically created from the routes array. Creates a Dashboard component that shows the Author's recently published. The profile sections shows the user profile image, name, and role along with recent activity on the user's post. The sidebar is full of NavLinks, when active, the main sections matches the current url and renders the correct component. The create article button brings the user to the new article form page where only basic inputs have been created.