Feat: Connects ArticleFormSection to the backend
Executes handleFormSubmit when form is submitted. Appends values from data and appends them to the new FormData object. Uses the sendRequest function to send a POST request to the backend route that creates a new article. Refactors the inputs to have the correct values and names to correspond with the Article Schema/Model.