Feat: Add a preview component to ArticleFormSection's form element
Tracks the current state of the form and when a valid file is chosen by the user, the file's name, size is set to an object and a new FileReader object is created, setting the value. Styles the new image preview and give user ability to remove image, clearing all the current variables related to the image and sets to default or null.